Bug 515233: Widen JSTreeContext::flags to 32 bits. r=igor
All the bits in this uint16 field are currently in use.  Adding bits
for projects like strict mode entails relocating existing flags, which
is additional work.  Furthermore, it seems that this has already
inspired people to put flags in places they don't belong:
TSF_DESTRUCTURING is a JSTokenStream flag, but is only used by the
parser.

This patch widens the field to 32 bits, and adjusts JSFunctionBox and
a few other places to match.

We should really replace these all with bitfields.

UltraBlame original commit: ff066b02d040aee4d01feedfa141c057174301e8

+++ js/src/jsemit.h
@@ -162,7 +162,7 @@ struct JSStmtInfo {
 #endif
 
 struct JSTreeContext {              
-    uint16          flags;          
+    uint32          flags;          
     uint16          ngvars;         
     uint32          bodyid;         
     uint32          blockidGen;     
@@ -238,11 +238,6 @@ struct JSTreeContext {
 
 
 #define TCF_RETURN_FLAGS        (TCF_RETURN_EXPR | TCF_RETURN_VOID)
-
-
-
-
-
 
 #define TCF_COMPILING           0x01 /* JSTreeContext is JSCodeGenerator */
 #define TCF_IN_FUNCTION         0x02 /* parsing inside function body */
